move batched pf
